filenames changed , heat map content styles
