--- src/components/heatMap/ChartScreen.tsx
@@ -1,232 +1,0 @@
-import * as React from 'react'
-import { Text, StyleSheet, FlatList, TouchableOpacity, ScrollView, View } from 'react-native'
-import { HeatMap } from './HeatMap'
-import { Col, Grid, Row } from 'native-base'
-import Accordion from 'react-native-collapsible/Accordion'
-import { useWindowSize } from './dynamicSize'
-
-interface ChartScreenProps {
-    sections: any[]
-    colors: string[]
-    // colorsPercentage: number[]
-    numberRange: number[]
-    //optionals
-    underlayColor?: String
-    expandMultiple?: boolean
-    xAxisOptions?: string[]
-}
-
-const defaultProps: ChartScreenProps = {
-    sections: [],
-    colors: ['#F0B22C', '#E77F24', '#E04931', '#732671'],
-    // colorsPercentage: [0, 25, 50, 75, 100],
-    numberRange: [0, 25, 50, 75, 100],
-    underlayColor: '#DEDEDE',
-    expandMultiple: false,
-    xAxisOptions: ['A', 'B', 'C', 'D', 'E'],
-}
-
-export const ChartScreen: React.FC<ChartScreenProps> = ({
-    sections,
-    colors,
-    // colorsPercentage,
-    underlayColor,
-    xAxisOptions,
-    numberRange,
-}) => {
-    const [activeSections, setActiveSections] = React.useState([])
-    const [colorsPercentage, setColorsPercentage] = React.useState<number[]>([])
-
-    const dynamicSize = useWindowSize()
-
-    const click = (value: number) => {
-        console.log('abcd', value)
-    }
-
-    React.useEffect(() => {
-        getColorPercentage()
-    }, [])
-
-    const getColorPercentage = () => {
-        let colorArray: number[] = []
-        let colorLength = colors.length
-        let colorSections = 100 / colorLength
-        console.log(colorSections, colorLength)
-        for (let i = 0; i <= colorLength; i++) {
-            colorArray.push(Math.round(colorSections * i))
-        }
-        setColorsPercentage(colorArray)
-    }
-
-    const _renderContent = (section: any) => {
-        return section && section.answer ? (
-            <View style={{ flex: 1 }}>
-                <Row style={styles.row}>
-                    <Col style={[styles.rowCol]}></Col>
-                    <Col style={styles.contentCol}>
-                        <Text style={{ textAlign: 'left', fontWeight: 'bold' }}>
-                            {'Question :'}
-                            {section.question}
-                        </Text>
-                        {section.answer.map((item: any, index: string) => {
-                            return <Text key={index}>{item.label}</Text>
-                        })}
-                    </Col>
-                </Row>
-            </View>
-        ) : (
-            <Text>No answers found</Text>
-        )
-    }
-
-    const _updateSections = (activeSections: any) => {
-        setActiveSections(activeSections)
-    }
-
-    const _renderHeader = (section: any) => {
-        return (
-            <View>
-                <Row style={styles.row}>
-                    <Col style={[styles.rowCol]}>
-                        <Text
-                            numberOfLines={2}
-                            style={{
-                                textAlign: 'center',
-                                justifyContent: 'center',
-                                flexWrap: 'wrap',
-                            }}
-                        >
-                            {section.question}
-                        </Text>
-                    </Col>
-                    <Col style={{ flex: 0.65, padding: 0 }}>
-                        <HeatMap
-                            numberOfLines={1}
-                            section={section}
-                            colors={colors}
-                            colorsPercentage={colorsPercentage}
-                            maximumValue={numberRange[numberRange.length - 1]}
-                            onBlockPress={value => click(value)}
-                        />
-                    </Col>
-                </Row>
-            </View>
-        )
-    }
-
-    const _getUnderlayColor = (): string => {
-        return underlayColor ? (underlayColor as string) : (defaultProps.underlayColor as string)
-    }
-
-    const generateNumberRange = () => {
-        let shiftedArray = numberRange ? [...numberRange] : [...colorsPercentage]
-        shiftedArray.shift()
-        return shiftedArray.map(item => {
-            return (
-                <Col key={Math.random()}>
-                    <Text style={{ textAlign: 'right' }}>{item}</Text>
-                </Col>
-            )
-        })
-    }
-
-    const generateColorRange = () => {
-        return colors.map(item => {
-            return (
-                <Col
-                    key={Math.random()}
-                    style={{ backgroundColor: item, flex: 1 / colors.length, height: 25 }}
-                ></Col>
-            )
-        })
-    }
-
-    const displayXAxisOptions = () => {
-        let xAxisLength = xAxisOptions?.length || 0
-        let options = xAxisLength > 0 ? xAxisOptions : defaultProps.xAxisOptions
-        return options?.map((item: any) => {
-            return (
-                <Col key={Math.random()} style={{ justifyContent: 'center' }}>
-                    <Text style={{ textAlign: 'center', marginVertical: 8 }}>{item}</Text>
-                </Col>
-            )
-        })
-    }
-
-    return (
-        <>
-            <View>
-                <Row style={[styles.row, { marginVertical: 10 }]}>
-                    <Col style={styles.rowCol}></Col>
-                    <Col style={styles.rangeCol}>
-                        <Row style={[{ width: dynamicSize.width * 0.5 }]}>
-                            <Text style={{ textAlign: 'left' }}>{colorsPercentage?.[0]}</Text>
-                            {generateNumberRange()}
-                        </Row>
-                        <Row style={{ width: dynamicSize.width * 0.5 }}>{generateColorRange()}</Row>
-                    </Col>
-                </Row>
-            </View>
-            <View style={{ paddingHorizontal: 20 }}>
-                <Row style={styles.row}>
-                    <Col style={styles.rowCol}></Col>
-                    <Col style={{ flex: 0.65, flexDirection: 'row', paddingRight: 5 }}>
-                        {displayXAxisOptions()}
-                    </Col>
-                </Row>
-            </View>
-            <Accordion
-                key={Math.random()}
-                sections={sections}
-                activeSections={activeSections}
-                renderHeader={_renderHeader}
-                renderContent={_renderContent}
-                onChange={_updateSections}
-                underlayColor={_getUnderlayColor()}
-                expandMultiple={true}
-                containerStyle={{ paddingHorizontal: 20, paddingBottom: 20 }}
-                duration={300}
-            />
-        </>
-    )
-}
-
-ChartScreen.defaultProps = defaultProps
-
-const styles = StyleSheet.create({
-    row: {
-        marginHorizontal: 5,
-        justifyContent: 'flex-end',
-        alignItems: 'flex-start',
-        width: '100%',
-    },
-    rowCol: {
-        flex: 0.35,
-        justifyContent: 'flex-end',
-        padding: 5,
-        // paddingRight:0,
-        flexWrap: 'wrap',
-    },
-    itemText: {
-        color: '#000',
-        marginLeft: 10,
-        margin: 5,
-        flex: 1,
-        flexWrap: 'nowrap',
-    },
-    contentCol: {
-        flex: 0.65,
-        marginLeft: 10,
-        marginVertical: 1,
-        padding: 8,
-        backgroundColor: '#DEDEDE',
-        borderRadius: 5,
-    },
-    rangeCol: {
-        flex: 0.65,
-        marginLeft: 10,
-        marginVertical: 1,
-        padding: 8,
-        borderRadius: 5,
-    },
-})
