+++ src/components/heatMap/index.ts
@@ -1,3 +1,3 @@
-export * from './ChartScreen'
-export * from './choicetype'
-export * from './dynamicSize'
+export * from './HeatMap'
+export * from './TransformReports'
+export * from './WindowSize'
